Optimized for MSVC support

+++ src/other_klvparser.c
@@ -10,7 +10,6 @@
 
 #include "include/Config.h"
 #include "libklv/libklv.h"
-#include "libklv/list.h"
 
 const size_t BYTES_IN_A_MEGABYTE = 1048576;
 
@@ -43,7 +42,7 @@ int main(int argc, char **argv) {
   if (binary) {
     klv_ctx_t *context = libklv_init();
     context->buffer = binary;
-    context->buffer_size = data_size;
+    context->buffer_size = (uint16_t)data_size;
     context->buf_ptr = binary;
     context->buf_end = &binary[data_size];
 
@@ -74,12 +73,12 @@ int read_data(uint8_t *buffer, FILE *in, size_t *size) {
       if (*size == UINT32_MAX) {
         fprintf(stderr, "Binary data loaded with size %llu\n", result);
       } else {
-        fprintf(stderr, "Only able to load %llu of %ld bytes\n", result, *size);
+        fprintf(stderr, "Only able to load %llu of %zu bytes\n", result, *size);
       }
-      *size = result;
+      *size = (size_t)result;
       return -1;
     } else {
-      fprintf(stderr, "Binary data loaded with size %ld\n", *size);
+      fprintf(stderr, "Binary data loaded with size %zu\n", *size);
     }
   }
